--- .eslintrc.cjs
@@ -1,25 +1,0 @@
-module.exports = {
-  root: true,
-  env: { browser: true, es2020: true },
-
-  extends: [
-    'eslint:recommended',
-    'plugin:@typescript-eslint/recommended-type-checked',
-    'plugin:react/recommended',
-    'plugin:react/jsx-runtime',
-    'plugin:react-hooks/recommended',
-    'plugin:prettier/recommended',
-  ],
-
-  settings: { react: { version: 'detect' } },
-
-  parser: '@typescript-eslint/parser',
-  parserOptions: {
-    ecmaVersion: 'latest',
-    sourceType: 'module',
-    project: ['./tsconfig.json', './tsconfig.node.json'],
-    tsconfigRootDir: __dirname,
-  },
-
-  plugins: ['react-refresh'],
-}
