+++ .husky/pre-commit
@@ -1,1 +1,3 @@
-bunx lint-staged
+bun biome check --write --no-errors-on-unmatched $(git diff --cached --name-only --diff-filter=ACMR)
+
+git update-index --again
