+++ .husky/pre-push
@@ -1,0 +1,1 @@
+bun run build
