+++ biome.json
@@ -1,0 +1,29 @@
+{
+   "$schema": "https://biomejs.dev/schemas/2.4.11/schema.json",
+
+   "files": {
+      "includes": ["**"]
+   },
+
+   "vcs": {
+      "enabled": true,
+      "clientKind": "git",
+      "useIgnoreFile": true
+   },
+
+   "formatter": {
+      "indentStyle": "space",
+      "indentWidth": 3,
+      "lineEnding": "lf"
+   },
+
+   "html": {
+      "experimentalFullSupportEnabled": true
+   },
+
+   "css": {
+      "parser": {
+         "tailwindDirectives": true
+      }
+   }
+}
