+++ package.json
@@ -1,48 +1,35 @@
 {
-  "name": "qr-bar-generator",
-  "private": true,
-  "version": "2.0.1",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build --emptyOutDir",
-    "lint:check": "eslint ./src/**/*.{ts,tsx} vite.config.ts",
-    "lint:fix": "eslint ./src/**/*.{ts,tsx} vite.config.ts --fix",
-    "preview": "vite preview",
-    "prepare": "husky"
-  },
-  "dependencies": {
-    "@radix-ui/themes": "^3.0.2",
-    "barcode-tool": "^1.1.1",
-    "lucide-react": "^0.367.0",
-    "qrcode": "^1.5.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "zustand": "^4.5.2"
-  },
-  "devDependencies": {
-    "@tailwindcss/vite": "^4.0.0-alpha.14",
-    "@types/qrcode": "^1.5.5",
-    "@types/react": "^18.2.66",
-    "@types/react-dom": "^18.2.22",
-    "@typescript-eslint/eslint-plugin": "^7.2.0",
-    "@typescript-eslint/parser": "^7.2.0",
-    "@vitejs/plugin-react-swc": "^3.5.0",
-    "eslint": "^8.57.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-react": "^7.34.1",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.6",
-    "husky": "^9.0.11",
-    "lint-staged": "^15.2.2",
-    "tailwindcss": "^4.0.0-alpha.14",
-    "typescript": "^5.2.2",
-    "vite": "^5.2.0"
-  },
-  "lint-staged": {
-    "*.{ts,tsx}": [
-      "eslint --fix --max-warnings 0"
-    ]
-  }
+   "name": "qr-bar-code-generator",
+   "private": true,
+   "version": "3.0.0",
+   "type": "module",
+   "packageManager": "bun@1.3.11",
+   "scripts": {
+      "dev": "vite",
+      "build": "tsc -b && vite build",
+      "preview": "vite preview",
+      "lint": "biome check",
+      "lint:fix": "biome check --write",
+      "prepare": "husky"
+   },
+   "dependencies": {
+      "@radix-ui/themes": "^3.3.0",
+      "jsbarcode": "^3.12.3",
+      "lucide-react": "^1.8.0",
+      "qrcode": "^1.5.4",
+      "react": "^19.2.4",
+      "react-dom": "^19.2.4"
+   },
+   "devDependencies": {
+      "@biomejs/biome": "2.4.11",
+      "@tailwindcss/vite": "^4.2.2",
+      "@types/qrcode": "^1.5.6",
+      "@types/react": "^19.2.14",
+      "@types/react-dom": "^19.2.3",
+      "@vitejs/plugin-react": "^6.0.1",
+      "husky": "^9.1.7",
+      "tailwindcss": "^4.2.2",
+      "typescript": "~6.0.2",
+      "vite": "^8.0.4"
+   }
 }
