--- src/index.css
@@ -1,6 +1,0 @@
-@import "tailwindcss";
-
-
-button:hover:not(:disabled) {
-  cursor: pointer;
-}
