+++ src/index.html
@@ -2,10 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <meta name="description" content="Browser extension for QR & Bar Code generation">
-    <script type="module" src="index.tsx"></script>
+    <title>qr-bar-code-generator</title>
   </head>
-
   <body>
+    <script type="module" src="./main.tsx"></script>
   </body>
 </html>
