+++ src/layout/qr-code-tab.tsx
@@ -1,34 +1,31 @@
-import { TabContent } from "~/components/tab-content";
-
 import QRCode from "qrcode";
-import { useState } from "react";
-import { SaveButton } from "~/components/save-button";
-import { CodeInput } from "~/components/code-input";
-import { CopyButton } from "~/components/copy-button";
+import { useRef } from "react";
+import { CodeInput } from "~/component/code-input";
+import { CopyButton } from "~/component/copy-button";
+import { SaveButton } from "~/component/save-button";
+import { TabContent } from "~/component/tab-content";
 
 function generateQR(value: string, canvas: HTMLCanvasElement | null) {
-  void QRCode.toCanvas(canvas, value);
+   QRCode.toCanvas(canvas, value);
 }
 
 export function QRCodeTab({ optionName }: { optionName: string }) {
-  const [canvasRef, setRef] = useState<HTMLCanvasElement | null>(null);
+   const canvasRef = useRef<HTMLCanvasElement | null>(null);
 
-  return (
-    <TabContent value={optionName}>
-      {canvasRef && (
-        <CodeInput
-          placeholder="Enter QR Code text"
-          generateFn={(value: string) => generateQR(value, canvasRef)}
-        />
-      )}
+   return (
+      <TabContent value={optionName}>
+         <CodeInput
+            placeholder="Enter QR Code text"
+            generateFn={(value: string) => generateQR(value, canvasRef.current)}
+         />
 
-      <canvas ref={(ref) => setRef(ref)} className="max-w-full" />
+         <canvas ref={canvasRef} className="max-w-full" />
 
-      <div className="flex gap-4 pt-4">
-        <SaveButton canvas={canvasRef} filename="qrcode.png" />
+         <div className="flex gap-4 pt-4">
+            <SaveButton canvas={canvasRef} filename="qrcode.png" />
 
-        <CopyButton canvas={canvasRef} />
-      </div>
-    </TabContent>
-  );
+            <CopyButton canvas={canvasRef} />
+         </div>
+      </TabContent>
+   );
 }
