--- src/lib/copy-image.ts
@@ -1,13 +1,0 @@
-export function copyImage(canvas: HTMLCanvasElement | null) {
-  if (!canvas) return;
-
-  canvas.toBlob((blob) => {
-    if (!blob) return;
-
-    void navigator.clipboard.write([
-      new ClipboardItem({
-        "image/png": blob,
-      }),
-    ]);
-  }, "image/png");
-}
