+++ src/main.css
@@ -1,0 +1,5 @@
+@import "tailwindcss";
+
+button:not(:disabled) {
+   cursor: pointer;
+}
